Simplify beatmap embed information
